+++ cspell.contributors.txt
@@ -38,4 +38,5 @@ janhassel
 bruells
 
 Jagath
+jagathgj
 Jayakumar
