+++ cspell.json
@@ -1,11 +1,7 @@
 {
   "version": "0.1",
   "language": "en",
-  "dictionaries": [
-    "contributors",
-    "html",
-    "packages"
-  ],
+  "dictionaries": ["contributors", "html", "packages"],
   "dictionaryDefinitions": [
     {
       "name": "contributors",
@@ -110,6 +106,7 @@
     "timeframe",
     "treeview",
     "uuidv",
-    "webfonts"
+    "webfonts",
+    "login"
   ]
 }
