+++ packages/cloud-cognitive/src/components/Datagrid/utils/getArgTypes.js
@@ -81,4 +81,10 @@ export const ARG_TYPES = {
     description:
       'This is an object containing all of the props used with the column customization extension. _This value is set/passed inside of the `datagridState` object._',
   },
+  filterProps: {
+    name: 'Filter props',
+    control: 'object',
+    description:
+      'This is an object for all the props passed into the filter flyout and filter panel',
+  },
 };
